a bit more code organization
