+++ cmd/agent-operator/README.md
@@ -3,8 +3,20 @@
 The Grafana Agent Operator is a Kubernetes operator that makes it easier to
 deploy Grafana Agent and easier to discover targets for metric collection.
 
-It is currently in **active development** and can't (and shouldn't!) be used in
-production environments.
+It is based on the [Prometheus Operator](https://github.com/prometheus-operator/prometheus-operator)
+and aims to be compatible the official ServiceMonitor, PodMonitor, and Probe
+CRDs that Prometheus Operator useres are used to.
+
+## Roadmap
+
+- [ ] Helm chart
+- [ ] Logs support
+- [ ] Traces support
+- [ ] Integrations support
+
+## Installing
+
+TODO. Stay tuned!
 
 ## Developer Reference
