+++ pkg/operator/reconciler.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"time"
 
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
@@ -15,7 +14,6 @@ import (
 	"github.com/grafana/agent/pkg/operator/clientutil"
 	"github.com/grafana/agent/pkg/operator/config"
 	"github.com/grafana/agent/pkg/operator/logutil"
-	prom "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	apps_v1 "k8s.io/api/apps/v1"
 	core_v1 "k8s.io/api/core/v1"
 	k8s_errors "k8s.io/apimachinery/pkg/api/errors"
@@ -69,189 +67,32 @@ func (r *reconciler) Reconcile(ctx context.Context, req controller.Request) (con
 		return controller.Result{}, nil
 	}
 
+	// Update our notifiers with the objects we discovered from building the
+	// deployment. This allows us to re-reconcile when any of the objects that
+	// composed our final deployment changes.
+	for _, secondary := range secondaryResources {
+		r.eventHandlers[secondary].Notify(req.NamespacedName, builder.ResourceSelectors[secondary])
+	}
+
 	// Fill secrets in store
 	if err := r.fillStore(ctx, deployment.AssetReferences(), secrets); err != nil {
 		level.Error(l).Log("msg", "unable to cache secrets for building config", "err", err)
 		return controller.Result{}, nil
 	}
 
-	// Create configuration in a secret
-	{
-		rawConfig, err := deployment.BuildConfig(secrets)
-
-		var jsonnetError jsonnet.RuntimeError
-		if errors.As(err, &jsonnetError) {
-			// Jump Jsonnet errors to the console to retain newlines and make them
-			// easier to digest.
-			fmt.Fprintf(os.Stderr, "%s", jsonnetError.Error())
-		}
+	type reconcileFunc func(context.Context, log.Logger, config.Deployment, assets.SecretStore) error
+	actors := []reconcileFunc{
+		r.createConfigurationSecret,
+		r.createSecrets,
+		r.createGoverningService,
+		r.createStatefulSets,
+	}
+	for _, actor := range actors {
+		err := actor(ctx, l, deployment, secrets)
 		if err != nil {
-			level.Error(l).Log("msg", "unable to build config", "err", err)
-			return controller.Result{}, nil
-		}
-
-		blockOwnerDeletion := true
-
-		secret := core_v1.Secret{
-			ObjectMeta: v1.ObjectMeta{
-				Namespace: agent.Namespace,
-				Name:      fmt.Sprintf("%s-config", agent.Name),
-				Labels:    r.config.Labels.Merge(managedByOperatorLabels),
-				OwnerReferences: []v1.OwnerReference{{
-					APIVersion:         agent.APIVersion,
-					BlockOwnerDeletion: &blockOwnerDeletion,
-					Kind:               agent.Kind,
-					Name:               agent.Name,
-					UID:                agent.UID,
-				}},
-			},
-			Data: map[string][]byte{"agent.yml": []byte(rawConfig)},
-		}
-
-		level.Info(l).Log("msg", "creating or updating secret", "secret", secret.Name)
-		err = r.Client.Update(ctx, &secret)
-		if k8s_errors.IsNotFound(err) {
-			err = r.Client.Create(ctx, &secret)
-		}
-		if k8s_errors.IsAlreadyExists(err) {
+			level.Error(l).Log("msg", "error during reconciling", "err", err)
 			return controller.Result{Requeue: true}, nil
 		}
-		if err != nil {
-			level.Error(l).Log("msg", "failed to create Secret for storing config", "err", err)
-			return controller.Result{Requeue: true}, nil
-		}
-	}
-
-	// Create secrets from asset store. These will be used to create volume
-	// mounts into pods.
-	{
-		blockOwnerDeletion := true
-
-		data := make(map[string][]byte)
-		for k, value := range secrets {
-			data[config.SanitizeLabelName(string(k))] = []byte(value)
-		}
-
-		secret := core_v1.Secret{
-			ObjectMeta: v1.ObjectMeta{
-				Namespace: agent.Namespace,
-				Name:      fmt.Sprintf("%s-secrets", agent.Name),
-				OwnerReferences: []v1.OwnerReference{{
-					APIVersion:         agent.APIVersion,
-					BlockOwnerDeletion: &blockOwnerDeletion,
-					Kind:               agent.Kind,
-					Name:               agent.Name,
-					UID:                agent.UID,
-				}},
-			},
-			Data: data,
-		}
-
-		level.Info(l).Log("msg", "creating or updating secret", "secret", secret.Name)
-		err = r.Client.Update(ctx, &secret)
-		if k8s_errors.IsNotFound(err) {
-			err = r.Client.Create(ctx, &secret)
-		}
-		if k8s_errors.IsAlreadyExists(err) {
-			return controller.Result{Requeue: true}, nil
-		}
-		if err != nil {
-			level.Error(l).Log("msg", "failed to create Secret for storing config", "err", err)
-			return controller.Result{Requeue: true}, nil
-		}
-	}
-
-	// Generate governing service.
-	{
-		svc := generateStatefulSetService(r.config, deployment)
-		level.Info(l).Log("msg", "creating or updating statefulset service", "service", svc.Name)
-		err := clientutil.CreateOrUpdateService(ctx, r.Client, svc)
-		if err != nil {
-			level.Error(l).Log("msg", "failed to create statefulset service", "err", err)
-			return controller.Result{}, nil
-		}
-	}
-
-	// Generate and create StatefulSet for all shards.
-	{
-		shards := minShards
-		if reqShards := deployment.Agent.Spec.Prometheus.Shards; reqShards != nil && *reqShards > 1 {
-			shards = *reqShards
-		}
-
-		// TODO(rfratto): when refactoring, keep in mind that returning an error
-		// will cause it to requeue.
-
-		// Keep track of generated stateful sets so we can delete ones that should
-		// no longer exist.
-		generated := make(map[string]struct{})
-
-		for shard := int32(0); shard < shards; shard++ {
-			name := deployment.Agent.Name
-			if shard > 0 {
-				name = fmt.Sprintf("%s-shard-%d", name, shard)
-			}
-
-			ss, err := generateStatefulSet(r.config, name, deployment, shard)
-			if err != nil {
-				level.Error(l).Log("msg", "failed to generate statefulset", "err", err)
-				return controller.Result{Requeue: false}, nil
-			}
-
-			level.Info(l).Log("msg", "creating or updating statefulset", "statefulset", ss.Name)
-			err = r.Client.Update(ctx, ss)
-			if k8s_errors.IsNotFound(err) {
-				err = r.Client.Create(ctx, ss)
-			}
-			if k8s_errors.IsNotAcceptable(err) {
-				level.Error(l).Log("msg", "unacceptable change to statefulset. deleting and re-querying", "err", err)
-				err = r.Client.Delete(ctx, ss)
-				if err != nil {
-					level.Error(l).Log("msg", "failed to delete unacceptable statefulset")
-				}
-				return controller.Result{
-					Requeue:      true,
-					RequeueAfter: 500 * time.Millisecond,
-				}, nil
-			}
-			if k8s_errors.IsAlreadyExists(err) {
-				return controller.Result{Requeue: true}, nil
-			}
-			if err != nil {
-				level.Error(l).Log("msg", "failed to create statefulset", "err", err)
-				return controller.Result{}, nil
-			}
-
-			generated[ss.Name] = struct{}{}
-		}
-
-		// Clean up statefulsets that should no longer exist.
-		var statefulSets apps_v1.StatefulSetList
-		err = r.List(ctx, &statefulSets, &client.ListOptions{
-			LabelSelector: labels.SelectorFromSet(labels.Set{
-				agentNameLabelName: deployment.Agent.Name,
-			}),
-		})
-		if err != nil {
-			level.Error(l).Log("msg", "failed to list statefulsets", "err", err)
-			return controller.Result{}, nil
-		}
-		for _, ss := range statefulSets.Items {
-			if _, keep := generated[ss.Name]; keep {
-				continue
-			}
-			level.Info(l).Log("msg", "deleting stale statefulset", "name", ss.Name)
-			if err := r.Client.Delete(ctx, &ss); err != nil {
-				level.Warn(l).Log("msg", "failed to delete stale statefulset", "err", err)
-			}
-		}
-	}
-
-	// Update our notifiers with every object we discovered. This ensures that we
-	// will re-reconcile whenever any of these objects (which composed our configs)
-	// changes.
-	for _, secondary := range secondaryResources {
-		r.eventHandlers[secondary].Notify(req.NamespacedName, builder.ResourceSelectors[secondary])
 	}
 
 	return controller.Result{}, nil
@@ -302,297 +143,163 @@ func (r *reconciler) fillStore(ctx context.Context, refs []config.AssetReference
 	return nil
 }
 
-type deploymentBuilder struct {
-	client.Client
-
-	Config  *Config
-	Agent   *grafana_v1alpha1.GrafanaAgent
-	Secrets assets.SecretStore
-
-	// ResourceSelectors is filled as objects are found and can be used to
-	// trigger future reconciles.
-	ResourceSelectors map[secondaryResource][]ResourceSelector
-}
-
-func (b *deploymentBuilder) Build(ctx context.Context, l log.Logger) (config.Deployment, error) {
-	instances, err := b.getPrometheusInstances(ctx)
-	if err != nil {
-		return config.Deployment{}, err
-	}
-	promInstances := make([]config.PrometheusInstance, 0, len(instances))
-
-	for _, inst := range instances {
-		sMons, err := b.getServiceMonitors(ctx, l, inst)
-		if err != nil {
-			return config.Deployment{}, fmt.Errorf("unable to fetch ServiceMonitors: %w", err)
-		}
-		pMons, err := b.getPodMonitors(ctx, inst)
-		if err != nil {
-			return config.Deployment{}, fmt.Errorf("unable to fetch PodMonitors: %w", err)
-		}
-		probes, err := b.getProbes(ctx, inst)
-		if err != nil {
-			return config.Deployment{}, fmt.Errorf("unable to fetch Probes: %w", err)
-		}
-
-		promInstances = append(promInstances, config.PrometheusInstance{
-			Instance:        inst,
-			ServiceMonitors: sMons,
-			PodMonitors:     pMons,
-			Probes:          probes,
-		})
-	}
-
-	return config.Deployment{
-		Agent:      b.Agent,
-		Prometheis: promInstances,
-	}, nil
-}
-
-func (b *deploymentBuilder) getPrometheusInstances(ctx context.Context) ([]*grafana_v1alpha1.PrometheusInstance, error) {
-	sel, err := b.getResourceSelector(
-		b.Agent.Namespace,
-		b.Agent.Spec.Prometheus.InstanceNamespaceSelector,
-		b.Agent.Spec.Prometheus.InstanceSelector,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("unable to build prometheus resource selector: %w", err)
-	}
-	b.ResourceSelectors[resourcePromInstance] = append(b.ResourceSelectors[resourcePromInstance], sel)
-
-	var (
-		list        grafana_v1alpha1.PrometheusInstanceList
-		namespace   = namespaceFromSelector(sel)
-		listOptions = &client.ListOptions{LabelSelector: sel.Labels, Namespace: namespace}
-	)
-	if err := b.List(ctx, &list, listOptions); err != nil {
-		return nil, err
-	}
-
-	items := make([]*grafana_v1alpha1.PrometheusInstance, 0, len(list.Items))
-	for _, item := range list.Items {
-		if match, err := b.matchNamespace(ctx, &item.ObjectMeta, sel); match {
-			items = append(items, item)
-		} else if err != nil {
-			return nil, fmt.Errorf("failed getting namespace: %w", err)
-		}
-	}
-	return items, nil
-}
-
-func (b *deploymentBuilder) getResourceSelector(
-	currentNamespace string,
-	namespaceSelector *v1.LabelSelector,
-	objectSelector *v1.LabelSelector,
-) (sel ResourceSelector, err error) {
-
-	// Set up our namespace label and object label selectors. By default, we'll
-	// match everything (the inverse of the k8s default). If we specify anything,
-	// we'll narrow it down.
-	var (
-		nsLabels  = labels.Everything()
-		objLabels = labels.Everything()
-	)
-	if namespaceSelector != nil {
-		nsLabels, err = v1.LabelSelectorAsSelector(namespaceSelector)
-		if err != nil {
-			return sel, err
-		}
-	}
-	if objectSelector != nil {
-		objLabels, err = v1.LabelSelectorAsSelector(objectSelector)
-		if err != nil {
-			return sel, err
-		}
-	}
-
-	sel = ResourceSelector{
-		NamespaceName: prom.NamespaceSelector{
-			MatchNames: []string{currentNamespace},
-		},
-		NamespaceLabels: nsLabels,
-		Labels:          objLabels,
-	}
-
-	// If we have a namespace selector, that means we're matching more than one
-	// namespace and we should adjust NamespaceName appropriatel.
-	if namespaceSelector != nil {
-		sel.NamespaceName = prom.NamespaceSelector{Any: true}
-	}
-
-	return
-}
-
-// namespaceFromSelector returns the namespace string that should be used for
-// querying lists of objects. If the ResourceSelector is looking at more than
-// one namespace, an empty string will be returned. Otherwise, it will return
-// the first namespace.
-func namespaceFromSelector(sel ResourceSelector) string {
-	if !sel.NamespaceName.Any && len(sel.NamespaceName.MatchNames) == 1 {
-		return sel.NamespaceName.MatchNames[0]
-	}
-	return ""
-}
-
-func (b *deploymentBuilder) matchNamespace(
-	ctx context.Context,
-	obj *v1.ObjectMeta,
-	sel ResourceSelector,
-) (bool, error) {
-	// If we were matching on a specific namespace, there's no
-	// further work to do here.
-	if namespaceFromSelector(sel) != "" {
-		return true, nil
-	}
-
-	var ns core_v1.Namespace
-	if err := b.Get(ctx, types.NamespacedName{Name: obj.Namespace}, &ns); err != nil {
-		return false, fmt.Errorf("failed getting namespace: %w", err)
-	}
-
-	return sel.NamespaceLabels.Matches(labels.Set(ns.Labels)), nil
-}
-
-func (b *deploymentBuilder) getServiceMonitors(
+// createConfigurationSecret creates the Grafana Agent configuration and stores
+// it into a secret.
+func (r *reconciler) createConfigurationSecret(
 	ctx context.Context,
 	l log.Logger,
-	inst *grafana_v1alpha1.PrometheusInstance,
-) ([]*prom.ServiceMonitor, error) {
-	sel, err := b.getResourceSelector(
-		inst.Namespace,
-		inst.Spec.ServiceMonitorNamespaceSelector,
-		inst.Spec.ServiceMonitorSelector,
-	)
+	d config.Deployment,
+	s assets.SecretStore,
+) error {
+
+	rawConfig, err := d.BuildConfig(s)
+
+	var jsonnetError jsonnet.RuntimeError
+	if errors.As(err, &jsonnetError) {
+		// Dump Jsonnet errors to the console to retain newlines and make them
+		// easier to digest.
+		fmt.Fprintf(os.Stderr, "%s", jsonnetError.Error())
+	}
 	if err != nil {
-		return nil, fmt.Errorf("unable to build service monitor resource selector: %w", err)
-	}
-	b.ResourceSelectors[resourceServiceMonitor] = append(b.ResourceSelectors[resourceServiceMonitor], sel)
-
-	var (
-		list        prom.ServiceMonitorList
-		namespace   = namespaceFromSelector(sel)
-		listOptions = &client.ListOptions{LabelSelector: sel.Labels, Namespace: namespace}
-	)
-	if err := b.List(ctx, &list, listOptions); err != nil {
-		return nil, err
+		return fmt.Errorf("unable to build config: %w", err)
 	}
 
-	items := make([]*prom.ServiceMonitor, 0, len(list.Items))
+	blockOwnerDeletion := true
 
-Item:
-	for _, item := range list.Items {
-		match, err := b.matchNamespace(ctx, &item.ObjectMeta, sel)
-		if err != nil {
-			return nil, fmt.Errorf("failed getting namespace: %w", err)
-		} else if !match {
-			continue
-		}
-
-		if b.Agent.Spec.Prometheus.ArbitraryFSAccessThroughSMs.Deny {
-			for _, ep := range item.Spec.Endpoints {
-				err := testForArbitraryFSAccess(ep)
-				if err == nil {
-					continue
-				}
-
-				level.Warn(l).Log(
-					"msg", "skipping service monitor",
-					"agent", client.ObjectKeyFromObject(b.Agent),
-					"servicemonitor", client.ObjectKeyFromObject(item),
-					"err", err,
-				)
-
-				continue Item
-			}
-		}
-
-		items = append(items, item)
-	}
-	return items, nil
-}
-
-func testForArbitraryFSAccess(e prom.Endpoint) error {
-	if e.BearerTokenFile != "" {
-		return fmt.Errorf("it accesses file system via bearer token file which is disallowed via GrafanaAgent specification")
+	secret := core_v1.Secret{
+		ObjectMeta: v1.ObjectMeta{
+			Namespace: d.Agent.Namespace,
+			Name:      fmt.Sprintf("%s-config", d.Agent.Name),
+			Labels:    r.config.Labels.Merge(managedByOperatorLabels),
+			OwnerReferences: []v1.OwnerReference{{
+				APIVersion:         d.Agent.APIVersion,
+				BlockOwnerDeletion: &blockOwnerDeletion,
+				Kind:               d.Agent.Kind,
+				Name:               d.Agent.Name,
+				UID:                d.Agent.UID,
+			}},
+		},
+		Data: map[string][]byte{"agent.yml": []byte(rawConfig)},
 	}
 
-	if e.TLSConfig == nil {
-		return nil
+	level.Info(l).Log("msg", "reconciling secret", "secret", secret.Name)
+	err = clientutil.CreateOrUpdateSecret(ctx, r.Client, &secret)
+	if err != nil {
+		return fmt.Errorf("failed to reconcile secret: %w", err)
 	}
-
-	if e.TLSConfig.CAFile != "" || e.TLSConfig.CertFile != "" || e.TLSConfig.KeyFile != "" {
-		return fmt.Errorf("it accesses file system via TLS config which is disallowed via GrafanaAgent specification")
-	}
-
 	return nil
 }
 
-func (b *deploymentBuilder) getPodMonitors(
+// createSecrets creates secrets from the secret store.
+func (r *reconciler) createSecrets(
 	ctx context.Context,
-	inst *grafana_v1alpha1.PrometheusInstance,
-) ([]*prom.PodMonitor, error) {
-	sel, err := b.getResourceSelector(
-		inst.Namespace,
-		inst.Spec.PodMonitorNamespaceSelector,
-		inst.Spec.PodMonitorSelector,
-	)
+	l log.Logger,
+	d config.Deployment,
+	s assets.SecretStore,
+) error {
+
+	blockOwnerDeletion := true
+
+	data := make(map[string][]byte)
+	for k, value := range s {
+		data[config.SanitizeLabelName(string(k))] = []byte(value)
+	}
+
+	secret := core_v1.Secret{
+		ObjectMeta: v1.ObjectMeta{
+			Namespace: d.Agent.Namespace,
+			Name:      fmt.Sprintf("%s-secrets", d.Agent.Name),
+			OwnerReferences: []v1.OwnerReference{{
+				APIVersion:         d.Agent.APIVersion,
+				BlockOwnerDeletion: &blockOwnerDeletion,
+				Kind:               d.Agent.Kind,
+				Name:               d.Agent.Name,
+				UID:                d.Agent.UID,
+			}},
+		},
+		Data: data,
+	}
+
+	level.Info(l).Log("msg", "reconciling secret", "secret", secret.Name)
+	err := clientutil.CreateOrUpdateSecret(ctx, r.Client, &secret)
 	if err != nil {
-		return nil, fmt.Errorf("unable to build service monitor resource selector: %w", err)
+		return fmt.Errorf("failed to reconcile secret: %w", err)
 	}
-	b.ResourceSelectors[resourcePodMonitor] = append(b.ResourceSelectors[resourcePodMonitor], sel)
-
-	var (
-		list        prom.PodMonitorList
-		namespace   = namespaceFromSelector(sel)
-		listOptions = &client.ListOptions{LabelSelector: sel.Labels, Namespace: namespace}
-	)
-	if err := b.List(ctx, &list, listOptions); err != nil {
-		return nil, err
-	}
-
-	items := make([]*prom.PodMonitor, 0, len(list.Items))
-	for _, item := range list.Items {
-		if match, err := b.matchNamespace(ctx, &item.ObjectMeta, sel); match {
-			items = append(items, item)
-		} else if err != nil {
-			return nil, fmt.Errorf("failed getting namespace: %w", err)
-		}
-	}
-	return items, nil
+	return nil
 }
 
-func (b *deploymentBuilder) getProbes(
+// createGoverningService creates the service that governs the (eventual)
+// StatefulSet. It must be created before the StatefulSet.
+func (r *reconciler) createGoverningService(
 	ctx context.Context,
-	inst *grafana_v1alpha1.PrometheusInstance,
-) ([]*prom.Probe, error) {
-	sel, err := b.getResourceSelector(
-		inst.Namespace,
-		inst.Spec.ProbeNamespaceSelector,
-		inst.Spec.ProbeSelector,
-	)
+	l log.Logger,
+	d config.Deployment,
+	s assets.SecretStore,
+) error {
+	svc := generateStatefulSetService(r.config, d)
+	level.Info(l).Log("msg", "reconciling statefulset service", "service", svc.Name)
+	err := clientutil.CreateOrUpdateService(ctx, r.Client, svc)
 	if err != nil {
-		return nil, fmt.Errorf("unable to build service monitor resource selector: %w", err)
+		return fmt.Errorf("failed to reconcile statefulset governing service: %w", err)
 	}
-	b.ResourceSelectors[resourceProbe] = append(b.ResourceSelectors[resourceProbe], sel)
+	return nil
+}
 
-	var namespace string
-	if !sel.NamespaceName.Any && len(sel.NamespaceName.MatchNames) == 1 {
-		namespace = sel.NamespaceName.MatchNames[0]
+// createStatefulSets creates a set of Grafana Agent StatefulSets, one per shard.
+func (r *reconciler) createStatefulSets(
+	ctx context.Context,
+	l log.Logger,
+	d config.Deployment,
+	s assets.SecretStore,
+) error {
+
+	shards := minShards
+	if reqShards := d.Agent.Spec.Prometheus.Shards; reqShards != nil && *reqShards > 1 {
+		shards = *reqShards
 	}
 
-	var list prom.ProbeList
-	listOptions := &client.ListOptions{LabelSelector: sel.Labels, Namespace: namespace}
-	if err := b.List(ctx, &list, listOptions); err != nil {
-		return nil, err
+	// Keep track of generated stateful sets so we can delete ones that should
+	// no longer exist.
+	generated := make(map[string]struct{})
+
+	for shard := int32(0); shard < shards; shard++ {
+		name := d.Agent.Name
+		if shard > 0 {
+			name = fmt.Sprintf("%s-shard-%d", name, shard)
+		}
+
+		ss, err := generateStatefulSet(r.config, name, d, shard)
+		if err != nil {
+			return fmt.Errorf("failed to generate statefulset for shard: %w", err)
+		}
+
+		level.Info(l).Log("msg", "reconciling statefulset", "statefulset", ss.Name)
+		err = clientutil.CreateOrUpdateStatefulSet(ctx, r.Client, ss)
+		if err != nil {
+			return fmt.Errorf("failed to reconcile statefulset for shard: %w", err)
+		}
+		generated[ss.Name] = struct{}{}
 	}
 
-	items := make([]*prom.Probe, 0, len(list.Items))
-	for _, item := range list.Items {
-		if match, err := b.matchNamespace(ctx, &item.ObjectMeta, sel); match {
-			items = append(items, item)
-		} else if err != nil {
-			return nil, fmt.Errorf("failed getting namespace: %w", err)
+	// Clean up statefulsets that should no longer exist.
+	var statefulSets apps_v1.StatefulSetList
+	err := r.List(ctx, &statefulSets, &client.ListOptions{
+		LabelSelector: labels.SelectorFromSet(labels.Set{
+			agentNameLabelName: d.Agent.Name,
+		}),
+	})
+	if err != nil {
+		return fmt.Errorf("failed to list statefulsets: %w", err)
+	}
+	for _, ss := range statefulSets.Items {
+		if _, keep := generated[ss.Name]; keep {
+			continue
+		}
+		level.Info(l).Log("msg", "deleting stale statefulset", "name", ss.Name)
+		if err := r.Client.Delete(ctx, &ss); err != nil {
+			return fmt.Errorf("failed to delete stale statefulset %s: %w", ss.Name, err)
 		}
 	}
-	return items, nil
+
+	return nil
 }
